Technically search works now.
I really have to fix my gitignore, too much garbage is getting committed.
Also, modified the activity to have action bars.  This breaks the detail
view.  Fix for that will be coming.
Search works with all 3 types of responses, people, movies, tv.
Had to mess with the model there to get that to work for all 3.

Left lots of todos with things to revisit in the future.
Big oofs.