Dagger2 and Caching integration
Important to note that Glide's cache is 250mb and has an automatic setting
of AUTOMATIC as of this commit. That means it'll store the transformed
image.

Integrated dagger (woo).  Might be wrong the way I do it, but it works
well and removed that stupid non-singleton retrofit thing I did.

That also allowed me to integrate okhttp caching, size of 10mb for now,
probably won't need larger since that data is updated daily anyway. Added
cache eviction to the only screen that currently supports refreshing,
popular movies.