Presenter clean up
As a stop gap, potentially, save the last item in the presenter
As to not exhaust my network call limit, just keep the last item hanging
around.  The data from the api is largely static and not going to change
fast enough that a new call needs to be made every time.  This will serve
as a (potentially) temporary solution.