Add Production Detail Page
This also includes adding a production detail item to the movie detail
state.
This was definitely questionable as there really is no need. The
production detail item contains a subset of information from the movie
detail, but as a stopgap, just do it this way. There's probably a better
way to map to those calls. The problem I ran into was mvrx doesn't allow
you to return from `withState` obviously and I am unsure of a way around
retrieving the latest state and mapping from that. Technically, it won't
be a problem since I only load once, but it'd bother me if I didn't do
it right.