Update landing page to include TV rows
This separates the previously triple bundled call into 5 separate
calls. Each one of these represents a row. Add 2 for tv.

Top rated seems like garbage, but I'll leave it for now and see
how I like it.

There are some UI updates here and each row now individually errors
and can be tapped for retry. There is also a loading animation for each.
I spoof the height here which results in a visual bug. Since there's no
good way of measuring views, I can only approximate row height with a
poster and so I use that to edit the layout params temporarily so that
loading doesn't make the view jump weirdly in size.
I made a todo about this and will one day either fix it or address
this whole situation differently.