UI Update
Fixed some bugs with the material themeing I was using. Status bar stays
as the accent blue green, but everything else is properly light or dark
themed otherwise. Replace all non-buttons with actual material buttons
by default, replace genre stuff with chips. Left a todo in there to add
for genre filters upon clicking on the chip. A lot of this makes a case
for view binding and compose, but we'll get there. This is the last UI
update before TV detail work begins.